action: Inline private/action.h into action.c
There is no need for a header since the types defined in it are
used only in action.c.
